Simplify making new eval environments for e-exps
